Rename the last of SunMiscSignalManger to just Signals
